quaintrade/requirements.txt
ConfigArgParse==1.7
dotmap==1.3.30
#fastapi
fyers_apiv3==3.1.1
jugaad_data==0.26
kiteconnect==5.0.1
lightweight_charts==2.0.1
matplotlib
mplfinance
numpy
pandas
pandas_ta==0.3.14b
pydantic==2.7.4
PyQt5==5.15.10
pytz==2024.1
PyYAML==6.0.1
schedule==1.2.2
scikit_learn==1.5.0
streamlit==1.36.0
streamlit_extras==0.4.3
streamlit_shortcuts==0.1.3
tabulate==0.9.0
